fix: all CI issues including lockfile, linting, and E2E tests
- Remove cmdk and update pnpm-lock.yaml.
- Remove unused SelectItem import in PresetSelector.tsx.
- Fix e2e/options_interaction.spec.ts to use correct locators and close dropdown after verification.
