+++ e2e/options_interaction.spec.ts
@@ -33,11 +33,16 @@ test("Options interaction behavior", async ({ page }) => {
 	await presetInput.press("Enter");
 
 	// ドロップダウンを開いて名前が反映されているか確認
-	await page.getByRole("button", { name: "プリセットを選択" }).click();
+	const selectButton = page.getByRole("combobox", { name: "プリセットを選択" });
+	await selectButton.click();
 	// ドロップダウン内の項目を特定するため、より具体的なロケータを使用（サイドバーにも同じテキストが表示されるため）
 	await expect(
-		page.getByRole("button", { name: "Test Preset" }).first(),
+		page.getByRole("option", { name: "Test Preset" }).first(),
 	).toBeVisible();
+
+	// 他の操作を妨げないようにドロップダウンを閉じる
+	await selectButton.click();
+	await expect(page.getByRole("option", { name: "Test Preset" })).not.toBeVisible();
 
 	// 別のカテゴリの操作を確認
 	const shuffleCategory = page
